+++ Data/Scripts/022_ML_AI/ML_Logger.rb
@@ -1,6 +1,6 @@
 #===============================================================================
 # a tiny static logger (singleton)
-# 
+#
 # Logs will be organized in ML/logs/battle[N]/turn[N]
 # for an example see ML/logs/battle0
 #===============================================================================
@@ -11,7 +11,9 @@ module ML_Logger
     @@battleID = 0
     @@turnID = 0
     @@situationID = 0
-    @@battlelogs = [] #currently not used (only required if ammended later on)
+    @@battleLogs = []
+    @@turnLogs = []
+    @@situationLogs = []
   
     #resets internal variables and creates directory
     #checks if the dir alrdy exist, if so calls itself recursivly to inc battleID (required after loading the game)
@@ -23,7 +25,9 @@ module ML_Logger
         self.incrBattleID
         battleLog = BattleLog.new(@@battleID, battle)
         battleLog.to_json(@@battleLogDir)
-        @@battlelogs[@@battleID] = battleLog
+        @@battleLogs[@@battleID] = battleLog
+        @@turnLogs = []
+        @@situationLogs = []
     end
 
     def self.incrBattleID
@@ -34,35 +38,22 @@ module ML_Logger
 
     #called in: pbOnAllBattlersEnteringBattle, pbEndOfRoundPhase
     def self.endTurn(battle)
-        turnLog = TurnLog.new(@@battleID, @@turnID, battle, @@battleLogDir, 1, "end")
+        turnLog = TurnLog.new(@@battleID, @@turnID, battle, @@battleLogDir, @@situationID, "end")
         #turnLog.to_json(@@battleLogDir+"/turn#{@@turnID}", "turn#{@@turnID}") #currently there is no need to store turn logs in a seperate folder
         turnLog.to_json(@@battleLogDir)
+        @@turnLogs[@@turnID] = turnLog
+        @@situationLogs = []
         @@turnID+=1
         @@situationID=0
-        # FOR TESTING   #TODO: remove!
-        #data = "test" #turnLog.to_s 
-        #result = `python ML/models/dummy/echo-score.py "#{data}"`.chomp
-        #echoln "AI would use the following move next: "
-        #echoln result #doesnt return anything (tbc)
-
-        #echoln "trying to make an HTTP call"
-        #ML_AI.testSocket #works
-
-        #echoln "trying to call AI through HTTP"
-        #ML_AI.runTest(); #https://stackoverflow.com/questions/76020053/tcpsocket-connection-to-azure-ml-endpoint 
-
-        result = `python "ML/models/maram-ml-kpwft/score.py"`.chomp
-        echoln "AI would use the following move next: "
-        echoln result
     end
 
     #called in pbEndOfBattle
     def self.endBattle(decision)
         if decision > 0
-            @@battlelogs[@@battleID].decision=decision
-            @@battlelogs[@@battleID].sid=1
-            @@battlelogs[@@battleID].sLabel="end"
-            @@battlelogs[@@battleID].to_json(@@battleLogDir)
+            @@battleLogs[@@battleID].decision=decision
+            @@battleLogs[@@battleID].sid=1
+            @@battleLogs[@@battleID].sLabel="end"
+            @@battleLogs[@@battleID].to_json(@@battleLogDir)
         end
     end
   
@@ -80,8 +71,8 @@ module ML_Logger
     #situations: "preSwitch", ?
     def self.newState(battle, situation)
         turnLog = TurnLog.new(@@battleID, @@turnID, battle, @@battleLogDir, @@situationID, situation)
-        #turnLog.to_json(@@battleLogDir+"/turn#{@@turnID}", "turn#{@@turnID}") #currently there is no need to store turn logs in a seperate folder
         turnLog.to_json(@@battleLogDir)
         @@situationID+=1
+        return turnLog
     end
 end